chore: extract ResourceFetcher
